+++ gi/pygi-cache.h
@@ -179,7 +179,7 @@ struct _PyGICallableCache {
     PyGIArgCache *user_data_varargs_arg;
 
     /* Number of args already added */
-    gssize args_offset;
+    unsigned int args_offset;
 
     /* Number of out args passed to gi_function_info_invoke.
      * This is used for the length of PyGIInvokeState.out_values */
